Move the findPath method to mounter, so that we can fake it properly
